Create custom dotcms hook and clean up
