+++ components/Loader.js
@@ -1,0 +1,6 @@
+/**
+ * Bare bones of this component, needs to expeand
+ */
+const Loader = () => <p>Loader...</p>;
+
+export default Loader;
